+++ lib/kandr-rails-extensions/array.rb
@@ -1,6 +1,9 @@
+# frozen_string_literal: true
+
 module KandrRailsExtensions
+  # Averaging functionality
   module ArrayAverage
-    def average(as_float=true)
+    def average(as_float = true)
       if as_float
         float_average
       else
@@ -16,6 +19,29 @@ module KandrRailsExtensions
       (sum / count.to_f) unless empty?
     end
   end
+
+  module ArrayIncludes
+    def self.included(base)
+      class << base
+        # Add in support for a more grammatical `includes?` method. Not using
+        # `alias` so that the method's existance and origin are more clear.
+        def includes?(item)
+          include?(item)
+        end
+      end
+
+      # Do the instance method too. Not using `alias` so that the method's
+      # existance and origin are more clear.
+      def includes?(item)
+        include?(item)
+      end
+    end
+  end
 end
 
-class Array; include KandrRailsExtensions::ArrayAverage; end
+
+# Add modifications to `Array` class
+class Array
+  include KandrRailsExtensions::ArrayAverage
+  include KandrRailsExtensions::ArrayIncludes
+end