+++ lib/kandr-rails-extensions/integer.rb
@@ -1,7 +1,14 @@
+# frozen_string_literal: true
 
 module KandrRailsExtensions
+  module IntegerBoolean
+    def to_bool
+      !zero?
+    end
+  end
+
+  # Ruby factorial functions
   module IntegerFactorial
-    # Ruby factorial function
     def factorial
       result = 1
       num = self
@@ -17,8 +24,8 @@ module KandrRailsExtensions
     end
   end
 
+  # Ruby termial functions
   module IntegerTermial
-    # Ruby termial function
     def termial
       result = 0
       num = self
@@ -35,7 +42,9 @@ module KandrRailsExtensions
   end
 end
 
+# Add modifications to `Integer` class
 class Integer
+  include KandrRailsExtensions::IntegerBoolean
   include KandrRailsExtensions::IntegerFactorial
   include KandrRailsExtensions::IntegerTermial
 end